requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
altair==5.0.1
    # via streamlit
attrs==23.1.0
    # via jsonschema
blinker==1.6.2
    # via streamlit
cachetools==5.3.1
    # via streamlit
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via streamlit
decorator==5.1.1
    # via validators
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via streamlit
greenlet==2.0.2
    # via sqlalchemy
idna==3.4
    # via requests
importlib-metadata==6.7.0
    # via streamlit
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonschema==4.17.3
    # via altair
loguru==0.7.0
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.25.0
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
packaging==23.1
    # via streamlit
pandas==2.0.3
    # via
    #   altair
    #   streamlit
pillow==9.5.0
    # via streamlit
protobuf==4.23.3
    # via streamlit
pyarrow==12.0.1
    # via streamlit
pycodestyle==2.10.0
    # via -r requirements.in
pydantic==1.10.11
    # via sqlmodel
pydeck==0.8.0
    # via streamlit
pygments==2.15.1
    # via rich
pympler==1.0.1
    # via streamlit
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   pandas
    #   streamlit
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
requests==2.31.0
    # via streamlit
rich==13.4.2
    # via streamlit
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
sqlalchemy==1.4.41
    # via sqlmodel
sqlalchemy2-stubs==0.0.2a34
    # via sqlmodel
sqlmodel==0.0.8
    # via -r requirements.in
streamlit==1.24.0
    # via -r requirements.in
tenacity==8.2.2
    # via streamlit
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.3.2
    # via streamlit
typing-extensions==4.7.1
    # via
    #   pydantic
    #   sqlalchemy2-stubs
    #   streamlit
tzdata==2023.3
    # via
    #   pandas
    #   pytz-deprecation-shim
tzlocal==4.3.1
    # via streamlit
urllib3==2.0.3
    # via requests
validators==0.20.0
    # via streamlit
watchdog==3.0.0
    # via streamlit
zipp==3.15.0
    # via importlib-metadata
